build: separate WMADEC from PLATFORM_WIN32
Add a new CMake option for WIN32 `WMA_DECODING` to build the WMA codec
decoding code, defaulting to `OFF`.

This makes it independent of `PLATFORM_WIN32` allowing it to be used
with SDL.
